Create Monty Hall Problem
Using a function simulate to count how often you succeed in the variable k. Pick a random door using the function randint. Write that code and flip your choice of the remaining door. It that remaining door is correct, imcrement k. Otherwise, don't increment k. When you run this 1000 times, the output should be approximately 2/3.